palmcash/requirements.txt
Django==4.2.7
django-widget-tweaks==1.4.12
sqlparse==0.4.4
xhtml2pdf==0.2.11
Pillow==10.0.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
reportlab==4.0.7
python-decouple==3.8
django-phonenumber-field==7.1.0
phonenumbers==8.13.25
mysqlclient==2.2.0

# Testing dependencies
hypothesis==6.92.1
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
